Desconecta sinais de indexação durante migração
