+++ sapl/norma/signals.py
@@ -1,27 +1,7 @@
-from subprocess import PIPE, call
-from threading import Thread
-
 from django.db.models.signals import post_delete, post_save
-
-from sapl.settings import PROJECT_DIR
+from sapl.utils import save_texto, delete_texto
 
 from .models import NormaJuridica
-
-
-class UpdateIndexCommand(Thread):
-    def run(self):
-        call([PROJECT_DIR.child('manage.py'), 'update_index'],
-             stdout=PIPE)
-
-
-def save_texto(sender, instance, **kwargs):
-    update_index = UpdateIndexCommand()
-    update_index.start()
-
-
-def delete_texto(sender, instance, **kwargs):
-    update_index = UpdateIndexCommand()
-    update_index.start()
 
 
 post_save.connect(save_texto, sender=NormaJuridica)
